% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qvd.R
\name{read_qvd}
\alias{read_qvd}
\title{Read qvd into R
Based on default path for 64bits qlivkiew. Designed for windows CMD.}
\usage{
read_qvd(qvd_path, qlikview_exe = "C:/Program Files/QlikView/Qv.EXE")
}
\arguments{
\item{qvd_path}{full qvd path (recommended the use together with here::here())}

\item{qlikview_exe}{full path to the executable qlikview file}
}
\value{
data frame
}
\description{
Read qvd into R
Based on default path for 64bits qlivkiew. Designed for windows CMD.
}
