% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload_qvw.R
\name{reload_qvw}
\alias{reload_qvw}
\title{Reload qlikview documents.
Based on default path for 32bits qlivkiew. Designed for windows CMD.
Useful to reload a qlikview document that exports csv, so that they can be refreshed to rerun an analysis based on latest data.}
\usage{
reload_qvw(path, qlikview_exe = "C:/Program Files/QlikView/Qv.EXE")
}
\arguments{
\item{path}{Path of qlikview document as an string (Function converts it to readable CMD format)}

\item{qlikview_exe}{full path to the executable qlikview file}
}
\value{
No output, just reloads the choosen qvw file
}
\description{
Reload qlikview documents.
Based on default path for 32bits qlivkiew. Designed for windows CMD.
Useful to reload a qlikview document that exports csv, so that they can be refreshed to rerun an analysis based on latest data.
}
